WIP: refactor accessor (@property) setup for Document/Case/Child.
