+++ src/fmu/sumo/explorer/objects/_child.py
@@ -1,27 +1,33 @@
 """module containing class for child object"""
-import re
+
 from typing import Dict
 from io import BytesIO
 from sumo.wrapper import SumoClient
 from fmu.sumo.explorer.objects._document import Document
 
-_prop_desc = [("name", "data.name", "Object name"),
-              ("dataname", "data.name", "Object name"),
-              ("classname", "class.name", "Object class name"),
-              ("casename", "fmu.case.name", "Object case name"),
-              ("content", "data.content", "Content"),
-              ("tagname", "data.tagname", "Object tagname"),
-              ("stratigraphic", "data.stratigraphic", "Object stratigraphic"),
-              ("vertical_domain", "data.vertical_domain", "Object vertical domain"),
-              ("context", "fmu.context.stage", "Object context"),
-              ("iteration", "fmu.iteration.name", "Object iteration"),
-              ("realization", "fmu.realization.id", "Object realization"),
-              ("aggregation", "fmu.aggregation.operation", "Object aggregation operation"),
-              ("stage", "fmu.context.stage", "Object stage"),
-              ("format", "data.format", "Object file format"),
-              ("dataformat", "data.format", "Object file format"),
-              ("relative_path", "file.relative_path", "Object relative file path")
-              ]
+_prop_desc = [
+    ("name", "data.name", "Object name"),
+    ("dataname", "data.name", "Object name"),
+    ("classname", "class.name", "Object class name"),
+    ("casename", "fmu.case.name", "Object case name"),
+    ("content", "data.content", "Content"),
+    ("tagname", "data.tagname", "Object tagname"),
+    ("stratigraphic", "data.stratigraphic", "Object stratigraphic"),
+    ("vertical_domain", "data.vertical_domain", "Object vertical domain"),
+    ("context", "fmu.context.stage", "Object context"),
+    ("iteration", "fmu.iteration.name", "Object iteration"),
+    ("realization", "fmu.realization.id", "Object realization"),
+    (
+        "aggregation",
+        "fmu.aggregation.operation",
+        "Object aggregation operation",
+    ),
+    ("stage", "fmu.context.stage", "Object stage"),
+    ("format", "data.format", "Object file format"),
+    ("dataformat", "data.format", "Object file format"),
+    ("relative_path", "file.relative_path", "Object relative file path"),
+]
+
 
 class Child(Document):
     """Class representing a child object in Sumo"""
@@ -54,14 +60,5 @@ class Child(Document):
 
         return self._blob
 
-_path_split_rx=re.compile("\]\.|\.|\[")
-def _splitpath(path):
-    parts = _path_split_rx.split(path)
-    return [int(x) if re.match("\d+", x) else x for x in parts]
 
-def _makeprop(attribute):
-    path = _splitpath(attribute)
-    return lambda self: self._get_property(path)
-
-for name, attribute, doc in _prop_desc:
-    setattr(Child, name, property(_makeprop(attribute), None, None, doc))
+Child.map_properties(Child, _prop_desc)
