+++ src/fmu/sumo/explorer/objects/_document.py
@@ -1,5 +1,19 @@
 """Contains class for one document"""
+
+import re
 from typing import List, Dict
+
+_path_split_rx = re.compile("\]\.|\.|\[")
+
+
+def _splitpath(path):
+    parts = _path_split_rx.split(path)
+    return [int(x) if re.match("\d+", x) else x for x in parts]
+
+
+def _makeprop(attribute):
+    path = _splitpath(attribute)
+    return lambda self: self._get_property(path)
 
 
 class Document:
@@ -40,3 +54,8 @@ class Document:
 
     def __getitem__(self, key: str):
         return self._metadata[key]
+
+    @staticmethod
+    def map_properties(cls, propmap):
+        for name, attribute, doc in propmap:
+            setattr(cls, name, property(_makeprop(attribute), None, None, doc))
